Fix bugs

+++ distributed_model.py
@@ -2,19 +2,20 @@ import torch
 import torch.nn as nn
 from transformer.encoder import TransformerEncoder
 
+
 class AE(nn.Module):
     def __init__(self, inp_dim, out_dim, cri='ce'):
         super().__init__()
 
         self.g = nn.Sequential(
             nn.Linear(inp_dim, out_dim),
-            nn.Tanh()            
+            nn.Tanh()
         )
 
         if cri == 'ce':
             self.h = nn.Sequential(
                 nn.Linear(out_dim, inp_dim),
-                nn.Tanh()            
+                nn.Tanh()
             )
         else:
             self.h = nn.Sequential(
@@ -26,7 +27,7 @@ class AE(nn.Module):
         else:
             self.cri = nn.MSELoss()
         self.mode = cri
-    
+
     def forward(self, x):
         enc_x = self.g(x)
         rec_x = self.h(enc_x)
@@ -34,6 +35,7 @@ class AE(nn.Module):
             return enc_x, self.cri(rec_x, x.argmax(1))
         else:
             return enc_x, self.cri(rec_x, x)
+
 
 class ENC(nn.Module):
     def __init__(self, inp_dim, out_dim, lab_dim=128, f='emb', n_heads=4, word_vec=None):
@@ -49,25 +51,27 @@ class ENC(nn.Module):
             if word_vec is not None:
                 self.f = nn.Embedding.from_pretrained(word_vec, freeze=False)
         elif f == 'lstm':
-            self.f = nn.LSTM(inp_dim, out_dim, bidirectional=True, batch_first=True)
+            self.f = nn.LSTM(inp_dim, out_dim,
+                             bidirectional=True, batch_first=True)
         elif f == 'trans':
-            self.f = TransformerEncoder(d_model=inp_dim, d_ff=out_dim, n_heads=n_heads)
+            self.f = TransformerEncoder(
+                d_model=inp_dim, d_ff=out_dim, n_heads=n_heads)
             self.b = nn.Sequential(
                 nn.Linear(inp_dim, lab_dim),
                 nn.Tanh()
             )
 
         self.cri = nn.MSELoss()
-    
+
     def forward(self, x, tgt, mask=None, h=None):
 
         if self.mode == 'emb':
-            enc_x = self.f(x)
+            enc_x = self.f(x.long())
         elif self.mode == 'lstm':
             enc_x, (h, c) = self.f(x, h)
         elif self.mode == 'trans':
             enc_x = self.f(x, mask=mask)
-        
+
         red_x = self.reduction(enc_x, mask, h)
         red_x = self.b(red_x)
         loss = self.cri(red_x, tgt)
@@ -89,6 +93,7 @@ class ENC(nn.Module):
             feat = torch.sum(x * mask.unsqueeze(-1), dim=1) / denom
             return feat
 
+
 class TransLayer(nn.Module):
     def __init__(self, inp_dim, lab_dim, hid_dim, lr):
         super().__init__()
@@ -98,44 +103,46 @@ class TransLayer(nn.Module):
 
         self.ae_opt = torch.optim.Adam(self.ae.parameters(), lr=0.0005)
         self.enc_opt = torch.optim.Adam(self.enc.parameters(), lr=lr)
-    
+
     def forward(self, x, y, mask):
 
         self.ae_opt.zero_grad()
-        enc_y , ae_loss = self.ae(y)
+        enc_y, ae_loss = self.ae(y)
         ae_loss.backward()
         self.ae_opt.step()
-    
+
         self.enc_opt.zero_grad()
         tgt = enc_y.clone().detach()
         enc_x, enc_loss, h, mask = self.enc(x, tgt, mask=mask)
         enc_loss.backward()
         self.enc_opt.step()
 
-        return enc_x.detach(), enc_y.detach(), ae_loss, enc_loss, mask        
+        return enc_x.detach(), enc_y.detach(), ae_loss, enc_loss, mask
 
 
 class TransModel(nn.Module):
     def __init__(self, vocab_size, emb_dim, l1_dim, lr, class_num, lab_dim=128, word_vec=None):
         super().__init__()
 
-        self.emb = EMBLayer(vocab_size, lab_dim, emb_dim, lr = 0.001, class_num=class_num, word_vec=word_vec)
+        self.emb = EMBLayer(vocab_size, lab_dim, emb_dim,
+                            lr=0.001, class_num=class_num, word_vec=word_vec)
         self.l1 = TransLayer(emb_dim, lab_dim, l1_dim, lr=lr)
         self.l1_dim = l1_dim
         self.l2 = TransLayer(emb_dim, lab_dim, l1_dim, lr=lr)
         self.losses = [0.0] * 6
         self.class_num = class_num
+
     def forward(self, x, y):
 
         mask = self.get_mask(x)
         y = torch.nn.functional.one_hot(y, self.class_num).float().to(y.device)
-        emb_x, emb_y, emb_ae, emb_as, _ = self.emb(x, y) # also updated
+        emb_x, emb_y, emb_ae, emb_as, _ = self.emb(x, y)  # also updated
 
         l1_x, l1_y, l1_ae, l1_as, mask = self.l1(emb_x, emb_y, mask)
         l2_x, l2_y, l2_ae, l2_as, mask = self.l2(l1_x, l1_y, mask)
 
         return [emb_ae, emb_as, l1_ae, l1_as, l2_ae, l2_as]
-    
+
     def get_mask(self, x):
         pad_mask = ~(x == 0)
         return pad_mask.cuda()
@@ -143,7 +150,7 @@ class TransModel(nn.Module):
     def inference(self, x):
 
         mask = self.get_mask(x)
-        emb_x = self.emb.enc.f(x)
+        emb_x = self.emb.enc.f(x.long())
         l1_x = self.l1.enc.f(emb_x, mask)
         l2_x = self.l2.enc.f(l1_x, mask)
 
@@ -157,23 +164,24 @@ class TransModel(nn.Module):
 
         return pred
 
+
 class LSTMLayer(nn.Module):
     def __init__(self, inp_dim, lab_dim, hid_dim, lr):
         super().__init__()
 
         self.enc = ENC(inp_dim, hid_dim, lab_dim=lab_dim, f='lstm')
         self.ae = AE(lab_dim, lab_dim, cri='mse')
-    
+
         self.ae_opt = torch.optim.Adam(self.ae.parameters(), lr=lr)
         self.enc_opt = torch.optim.Adam(self.enc.parameters(), lr=lr)
 
     def forward(self, x, y, mask=None, h=None):
 
         self.ae_opt.zero_grad()
-        enc_y , ae_loss = self.ae(y)
+        enc_y, ae_loss = self.ae(y)
         ae_loss.backward()
         self.ae_opt.step()
-    
+
         self.enc_opt.zero_grad()
         tgt = enc_y.clone().detach()
         enc_x, enc_loss, hidden, _ = self.enc(x, tgt, mask, h)
@@ -185,24 +193,26 @@ class LSTMLayer(nn.Module):
 
         return enc_x.detach(), enc_y.detach(), ae_loss, enc_loss, [hidden, mask]
 
+
 class EMBLayer(nn.Module):
     def __init__(self, inp_dim, lab_dim, hid_dim, lr, class_num=None, word_vec=None):
         super().__init__()
 
-        self.enc = ENC(inp_dim, hid_dim, lab_dim=lab_dim, f='emb', word_vec=word_vec)
+        self.enc = ENC(inp_dim, hid_dim, lab_dim=lab_dim,
+                       f='emb', word_vec=word_vec)
         assert class_num is not None
         self.ae = AE(class_num, lab_dim, cri='ce')
-    
+
         self.ae_opt = torch.optim.Adam(self.ae.parameters(), lr=lr)
         self.enc_opt = torch.optim.Adam(self.enc.parameters(), lr=lr)
 
     def forward(self, x, y, mask=None, h=None):
 
         self.ae_opt.zero_grad()
-        enc_y , ae_loss = self.ae(y)
+        enc_y, ae_loss = self.ae(y)
         ae_loss.backward()
         self.ae_opt.step()
-    
+
         self.enc_opt.zero_grad()
         tgt = enc_y.clone().detach()
         enc_x, enc_loss, hidden, mask = self.enc(x, tgt, mask, h)
@@ -211,11 +221,13 @@ class EMBLayer(nn.Module):
 
         return enc_x.detach(), enc_y.detach(), ae_loss, enc_loss, [hidden, mask]
 
+
 class LSTMModel(nn.Module):
     def __init__(self, vocab_size, emb_dim, l1_dim, lr, class_num, lab_dim=128, word_vec=None):
         super().__init__()
 
-        self.emb = EMBLayer(vocab_size, lab_dim, emb_dim, lr = 0.001, class_num=class_num, word_vec=word_vec)
+        self.emb = EMBLayer(vocab_size, lab_dim, emb_dim,
+                            lr=0.001, class_num=class_num, word_vec=word_vec)
         self.l1 = LSTMLayer(emb_dim, lab_dim, l1_dim, lr=lr)
         self.l1_dim = l1_dim
         self.l2 = LSTMLayer(l1_dim*2, lab_dim, l1_dim, lr=lr)
@@ -223,21 +235,23 @@ class LSTMModel(nn.Module):
         self.class_num = class_num
 
     def forward(self, x, y):
-        
+
         y = torch.nn.functional.one_hot(y, self.class_num).float().to(y.device)
-        emb_x, emb_y, emb_ae, emb_as, _ = self.emb(x, y) # also updated
+        emb_x, emb_y, emb_ae, emb_as, _ = self.emb(x, y)  # also updated
 
         l1_x, l1_y, l1_ae, l1_as, [h, _] = self.l1(emb_x, emb_y)
-        l1_x = torch.cat((l1_x[:, :, :self.l1_dim], l1_x[:, :, self.l1_dim:]), dim=-1)
+        l1_x = torch.cat(
+            (l1_x[:, :, :self.l1_dim], l1_x[:, :, self.l1_dim:]), dim=-1)
 
         l2_x, l2_y, l2_ae, l2_as, [h, _] = self.l2(l1_x, l1_y, h)
 
         return [emb_ae, emb_as, l1_ae, l1_as, l2_ae, l2_as]
-    
+
     def inference(self, x):
-        emb_x = self.emb.enc.f(x)
+        emb_x = self.emb.enc.f(x.long())
         l1_x, (h, c) = self.l1.enc.f(emb_x)
-        l1_x = torch.cat((l1_x[:, :, :self.l1_dim], l1_x[:, :, self.l1_dim:]), dim=-1)
+        l1_x = torch.cat(
+            (l1_x[:, :, :self.l1_dim], l1_x[:, :, self.l1_dim:]), dim=-1)
         # print(l1_x.shape, )
         l2_x, (h, c) = self.l2.enc.f(l1_x, (h, c))
         h = h[0] + h[1]
